# backend/schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Dagger API",
    "description": "API for Dagger AI application",
    "version": "0.0.1"
  },
  "paths": {
...
    "/dag/": {
      "get": {
        "tags": [
          "dag"
        ],
        "summary": "Get All Dags",
        "operationId": "get_all_dags_dag__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/dag"
                  },
                  "type": "array",
                  "title": "Response Get All Dags Dag  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "dag"
        ],
        "summary": "Dag Action",
        "operationId": "dag_action_dag__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DagRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DagResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/week/": {
      "get": {
        "tags": [
          "week"
        ],
        "summary": "Get Weeks",
        "operationId": "get_weeks_week__get",
        "parameters": [
          {
            "name": "request_type",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/WeekRequestType"
            }
          },
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Query"
            }
          },
          {
            "name": "week_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Week Id"
            }
          },
          {
            "name": "number_of_weeks",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Number Of Weeks"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          },
          {
            "name": "team_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Team Id"
            }
          },
          {
            "name": "collaborators",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Collaborators"
            }
          },
          {
            "name": "missed_deadlines_range",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "prefixItems": [
                    {
                      "type": "integer"
                    },
                    {
                      "type": "integer"
                    }
                  ],
                  "minItems": 2,
                  "maxItems": 2
                },
                {
                  "type": "null"
                }
              ],
              "title": "Missed Deadlines Range"
            }
          },
          {
            "name": "completed_task_range",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "prefixItems": [
                    {
                      "type": "integer"
                    },
                    {
                      "type": "integer"
                    }
                  ],
                  "minItems": 2,
                  "maxItems": 2
                },
                {
                  "type": "null"
                }
              ],
              "title": "Completed Task Range"
            }
          },
          {
            "name": "points_range",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "prefixItems": [
                    {
                      "type": "integer"
                    },
                    {
                      "type": "integer"
                    }
                  ],
                  "minItems": 2,
                  "maxItems": 2
                },
                {
                  "type": "null"
                }
              ],
              "title": "Points Range"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeekResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DagAction": {
        "type": "string",
        "enum": [
          "create",
          "add_edge",
          "delete_edge"
        ],
        "title": "DagAction"
      },
      "DagRequest": {
        "properties": {
          "dag_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dag Id"
          },
          "first_task_id": {
            "type": "string",
            "format": "uuid",
            "title": "First Task Id"
          },
          "second_task_id": {
            "type": "string",
            "format": "uuid",
            "title": "Second Task Id"
          },
          "team_id": {
            "type": "string",
            "format": "uuid",
            "title": "Team Id"
          },
          "action": {
            "$ref": "#/components/schemas/DagAction"
          }
        },
        "type": "object",
        "required": [
          "first_task_id",
          "team_id",
          "action"
        ],
        "title": "DagRequest"
      },
      "DagResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "dag_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dag Id"
          },
          "new_dag_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "New Dag Id"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "title": "DagResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WeekRequestType": {
        "type": "string",
        "enum": [
          "search_query",
          "compare_weeks",
          "get_weeks"
        ],
        "title": "WeekRequestType"
      },
      "WeekResponse": {
        "properties": {
          "weeks": {
            "items": {
              "$ref": "#/components/schemas/week"
            },
            "type": "array",
            "title": "Weeks"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          }
        },
        "type": "object",
        "required": [
          "weeks",
          "total_count"
        ],
        "title": "WeekResponse"
      },
      "dag": {
        "properties": {
          "dag_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dag Id"
          },
          "team_id": {
            "type": "string",
            "format": "uuid",
            "title": "Team Id"
          },
          "dag_graph": {
            "additionalProperties": {
              "items": {
                "type": "string",
                "format": "uuid"
              },
              "type": "array"
            },
            "propertyNames": {
              "format": "uuid"
            },
            "type": "object",
            "title": "Dag Graph"
          }
        },
        "type": "object",
        "required": [
          "team_id",
          "dag_graph"
        ],
        "title": "dag"
      },
...
      "week": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "start_date": {
            "type": "string",
            "format": "date-time",
            "title": "Start Date",
            "description": "The start date of the week"
          },
          "end_date": {
            "type": "string",
            "format": "date-time",
            "title": "End Date",
            "description": "The end date of the week (7 days after start_date)"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id",
            "description": "The ID of the user whose productivity is being summarized"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary",
            "description": "AI-generated summary of the week's work and productivity patterns"
          },
          "feedback": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback",
            "description": "AI-generated feedback aimed at improving worker productivity"
          },
          "collaborators": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "uuid"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collaborators",
            "description": "List of user IDs who were assigned roles on tasks during this week"
          },
          "missed_deadlines": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "uuid"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Missed Deadlines",
            "description": "List of task IDs where deadlines were missed during this week"
          },
          "completed_tasks": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "uuid"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed Tasks",
            "description": "List of task IDs that were completed during this week"
          },
          "points_completed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Points Completed",
            "description": "Total points completed, where points represent estimated hours of work"
          }
        },
        "type": "object",
        "required": [
          "start_date",
          "end_date",
          "user_id"
        ],
        "title": "week",
        "description": "A model representing a weekly summary of a user's productivity and accomplishments.\n\nThis model captures a week's worth of work activity, including tasks completed,\ncollaboration metrics, and AI-generated feedback to improve productivity.\nEach week spans from start_date to end_date (7 days later)."
      }
    }
  }
}